+++ src/hooks/reactQuery/auth/useLoginMutation.ts
@@ -1,7 +1,10 @@
 import { postLogin } from '@/apis/auth';
-import { setCookie } from '@/utils/cookie';
+import { getCookie, setCookie } from '@/utils/cookie';
+import jwtDecode from '@/utils/jwtDecode';
+import { useUserStore } from '@/utils/zustand';
 import { useMutation } from '@tanstack/react-query';
 import axios from 'axios';
+import { useNavigate } from 'react-router-dom';
 
 type LoginForm = {
   email: string;
@@ -9,12 +12,22 @@ type LoginForm = {
 };
 
 const useLoginMutation = () => {
+  const setUserInfo = useUserStore((state) => state.setUserInfo);
+  const navigate = useNavigate();
   return useMutation({
     mutationFn: async (data: LoginForm) => {
       const res = await postLogin(data);
       const ACCESS_TOKEN = res.data.accessToken;
       setCookie('accessToken', ACCESS_TOKEN);
       return res.data;
+    },
+    onSuccess: () => {
+      const token = getCookie('accessToken');
+      if (token) {
+        const userInfo = jwtDecode(token);
+        setUserInfo(userInfo);
+        navigate('/');
+      }
     },
     onError(error) {
       if (axios.isAxiosError(error) && error.response?.status === 401) {
